# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "GitHubSettings": {
        "additionalProperties": false,
        "description": "What the settings page reads and writes. The token itself is never returned.",
        "properties": {
          "base_branch": {
            "default": "main",
            "title": "Base Branch",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default owner/org for new projects.",
            "title": "Owner"
          },
          "token_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last four characters, if set.",
            "title": "Token Hint"
          },
          "token_set": {
            "default": false,
            "title": "Token Set",
            "type": "boolean"
          }
        },
        "required": [
          "base_branch",
          "owner",
          "token_hint",
          "token_set"
        ],
        "title": "GitHubSettings",
        "type": "object"
      },
      "GitHubSettingsIn": {
        "properties": {
          "base_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Branch"
          },
          "clear_token": {
            "default": false,
            "title": "Clear Token",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to keep the stored token.",
            "title": "Token"
          }
        },
        "title": "GitHubSettingsIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Slipwright",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings/github": {
      "get": {
        "operationId": "get_github_api_settings_github_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Github",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_github_api_settings_github_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Github",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/settings/jira": {
      "get": {
        "operationId": "get_jira_api_settings_jira_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Jira",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_jira_api_settings_jira_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JiraSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JiraSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Jira",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/settings/notifications": {
      "get": {
        "operationId": "get_notifications_api_settings_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Notifications",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_notifications_api_settings_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Notifications",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/settings/providers/{name}": {
      "put": {
        "operationId": "put_provider_api_settings_providers__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderSettings"
                  },
                  "title": "Response Put Provider Api Settings Providers  Name  Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Provider",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/providers/{name}/models": {
      "get": {
        "operationId": "provider_models_api_settings_providers__name__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModels"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Models",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/providers/{name}/test": {
      "post": {
        "description": "Call the vendor with the stored key and return the models it can use.",
        "operationId": "test_provider_api_settings_providers__name__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModels"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Provider",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/standards": {
      "get": {
        "operationId": "get_standards_api_settings_standards_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardsStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Standards",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_standards_api_settings_standards_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardsSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardsSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Standards",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/standards/reindex": {
      "post": {
        "operationId": "reindex_standards_api_settings_standards_reindex_post",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardsStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reindex Standards",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/standards/search": {
      "get": {
        "description": "Try a query the way the agents do: ranked sections with their scores.",
        "operationId": "search_standards_api_settings_standards_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "default": 4,
              "title": "K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StandardsHit"
                  },
                  "title": "Response Search Standards Api Settings Standards Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Standards",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/standards/pages": {
      "get": {
        "description": "The pages of a domain (plus ``core.md``) in the global corpus or a project's overrides.",
        "operationId": "list_standards_pages_api_standards_pages_get",
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StandardsPage"
                  },
                  "title": "Response List Standards Pages Api Standards Pages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Standards Pages",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_standards_page_api_standards_pages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Standards Page",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}